Fix - Refactor code in order to remove rewrite and only extend methods of CatalogSearch Collection
