+++ etc/adminhtml/di.xml
@@ -27,6 +27,4 @@
     <type name="Magento\Catalog\Model\Product">
         <plugin name="cacheClean" type="Smile\CustomEntityProductLink\Plugin\Model\ProductPlugin" sortOrder="10" />
     </type>
-
-    <preference for="Magento\CatalogSearch\Model\ResourceModel\Search\Collection" type="Smile\CustomEntityProductLink\Rewrite\Model\ResourceModel\Search\CustomCollection" />
 </config>
